msm: kgsl: Split out the context and command batch flags
Split out the context and command batch flags into separate piles.
In reality it is more like a venn diagram with some context
exlusive bits some cmdbatch exclusive bits and some shared bits.

The object of all this is to increase the available pool of exclusive
command batch flags and keep from using up all of the context bits
too. We still have to be careful for future shared bits so not all of
the unused field is up for grabs for the command batch but we
certainly have increased our options.

Change-Id: Ic0dedbadac4a12b712c911760f83da3436550bcb
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Sunil Khatri <sunilkh@codeaurora.org>
Signed-off-by: Konstantinos <kjpliakas@gmail.com>